+++ llvm/utils/TableGen/Common/CMakeLists.txt
@@ -27,7 +27,6 @@ add_llvm_library(LLVMTableGenCommon STATIC OBJECT EXCLUDE_FROM_ALL DISABLE_LLVM_
   CodeGenRegisters.cpp
   CodeGenSchedule.cpp
   CodeGenTarget.cpp
-  DAGISelMatcher.cpp
   InfoByHwMode.cpp
   InstructionEncoding.cpp
   OptEmitter.cpp
